requirements.txt
discord
google-cloud-firestore